删除无用代码，清理项目结构
